kotopedia-django/requirements.txt
django
djangorestframework
django-cors-headers
Pillow
django-choices
